New feature: renumber the pages of a PDF
Often, articles and numerized books have their own page numbering that
doesn't match the one of the PDF. That is, the first page of the book is
not at the first page of the PDF, but after.

The PDF file format allows to specify custom page numbers, but few PDF
use this feature.

This commit adds a feature that allows easier navigation in such PDFs.
The user should go to the first page of the book inside the PDF, and
click "Mark this page as page №1" (in the secondary menu bar). pdf.js
will then renumber all the pages in the pdf, so that pages before the
current page have a page label indicationg "intro-n" and pages starting
from the current one will have numbers corresponding to their real
position inside the book.

Signed-off-by: @lovasoa